+++ modules/extra/index.mjs
@@ -5,3 +5,4 @@
 import './GSChart.mjs';
 import './GSHighlight.mjs';
 import './GSMonaco.mjs';
+import './GSMarkdown.mjs';